refactor: consolidate context window tables into cost.ts
